Refactor tests
Move helper code to a separate file.
