+++ cas/filesystem_test.go
@@ -1,16 +1,10 @@
 package cas
 
 import (
-	"bytes"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
-)
-
-const (
-	filesystemTestBlobName = "ZZ8FaUwURAkWvzbnRhTt2pWSJCYZMAELqPk9USTUJgC4"
 )
 
 func temporaryFS() (*fileSystem, func()) {
@@ -37,21 +31,17 @@ func protect(fName string) func() {
 	return func() { os.Chmod(fName, mode) }
 }
 
-func testData() io.ReadCloser {
-	return ioutil.NopCloser(bytes.NewBuffer([]byte{}))
-}
-
 func TestFilesystemOpenFailure(t *testing.T) {
 
 	fs, d := temporaryFS()
 	defer d()
 
 	// Put file where directory should be
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	errPanic(err)
 	defer protect(touchFile(fName))()
 
-	stream, err := fs.Open(filesystemTestBlobName)
+	stream, err := fs.Open(emptyBlobName)
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -69,12 +59,12 @@ func TestFilesystemSaveFailureDir(t *testing.T) {
 	defer d()
 
 	// Put file where directory should be
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	errPanic(err)
 	fName = filepath.Dir(fName)
 	defer protect(touchFile(fName))()
 
-	err = fs.Save(filesystemTestBlobName, testData())
+	err = fs.Save(emptyBlobName, emptyBlobReader())
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -88,12 +78,12 @@ func TestFilesystemToManySimultaneousSaves(t *testing.T) {
 	fs, d := temporaryFS()
 	defer d()
 
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	errPanic(err)
 	for i := 0; i < fileSystemMaxSimultaneousUploads; i++ {
 		touchFile(fs.temporaryWriteStreamFileName(fName, i))
 	}
-	err = fs.Save(filesystemTestBlobName, testData())
+	err = fs.Save(emptyBlobName, emptyBlobReader())
 	if err != errToManySimultaneousUploads {
 		t.Fatalf("Incorrect error received: %v", err)
 	}
@@ -106,13 +96,13 @@ func TestFilesystemSaveFailureTempFile(t *testing.T) {
 	defer d()
 
 	// Create blob's directory as unmodifiable
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	errPanic(err)
 	dirPath := filepath.Dir(fName)
 	errPanic(os.MkdirAll(dirPath, 0777))
 	defer protect(dirPath)()
 
-	err = fs.Save(filesystemTestBlobName, testData())
+	err = fs.Save(emptyBlobName, emptyBlobReader())
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -131,7 +121,7 @@ func TestFilesystemSaveAutoNamedFailureTempFile(t *testing.T) {
 	errPanic(os.MkdirAll(dirPath, 0777))
 	defer protect(dirPath)()
 
-	name, err := fs.SaveAutoNamed(testData())
+	name, err := fs.SaveAutoNamed(emptyBlobReader())
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -149,10 +139,10 @@ func TestFilesystemRenameFailure(t *testing.T) {
 	defer d()
 
 	// Create directory where blob should be
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	os.MkdirAll(fName, 0777)
 
-	err = fs.Save(filesystemTestBlobName, testData())
+	err = fs.Save(emptyBlobName, emptyBlobReader())
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -167,11 +157,11 @@ func TestFilesystemDeleteFailure(t *testing.T) {
 	defer d()
 
 	// Create directory where blob should be with some file inside
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	os.MkdirAll(fName, 0777)
 	touchFile(fName + "/keep.me")
 
-	err = fs.Delete(filesystemTestBlobName)
+	err = fs.Delete(emptyBlobName)
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
@@ -185,7 +175,7 @@ func TestFilesystemDeleteNotFound(t *testing.T) {
 	fs, d := temporaryFS()
 	defer d()
 
-	err := fs.Delete(filesystemTestBlobName)
+	err := fs.Delete(emptyBlobName)
 	if err != ErrNotFound {
 		t.Fatalf("Incorrect error received: %v", err)
 	}
@@ -197,13 +187,13 @@ func TestFilesystemExistsFailure(t *testing.T) {
 	defer d()
 
 	// Create blob's directory as unmodifiable
-	fName, err := fs.getFileName(filesystemTestBlobName)
+	fName, err := fs.getFileName(emptyBlobName)
 	errPanic(err)
 	dirPath := filepath.Dir(fName)
 	errPanic(os.MkdirAll(dirPath, 0777))
 	defer protect(dirPath)()
 
-	_, err = fs.Exists(filesystemTestBlobName)
+	_, err = fs.Exists(emptyBlobName)
 	if err == nil {
 		t.Fatal("Did not get error while trying to save blob")
 	}
